maxim-backend-geem/requirements.txt
starlette
uvicorn
tortoise-orm
asyncpg
marshmallow
boto3
